lab3/gcfun3/requirements.txt
pandas
keras
numpy
tensorflow
h5py
google-cloud-storage
google-cloud-pubsub